+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # 0.4.11 (2020-11-XX)
 
-  * Nelson Engine API for C (compatible with MEX Engine).
+  * Nelson Engine API for C (compatible with MEX Engine 100%).
     - engSetVisible,
-    - engGetVisible.
+    - engGetVisible,
+    - engEvalString,
+    - engOutputBuffer.
 
   * ```sha256``` builtin: get sha256 checksum of a file or a string.